apps/api/pyproject.toml
[project]
name = "docscribe-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "anthropic>=0.40.0",
    "groq>=0.13.0",
    "PyJWT>=2.10.0",
    "python-multipart>=0.0.12",
    "reportlab>=4.0.0",
    "alembic>=1.14.0",
]
